refactor(brain): split the 2.6k-line metrics.rs by metric family
metrics.rs becomes metrics/mod.rs (read_all_summaries, dispatch, print_help,
print_scorecard, shared render helpers) plus one submodule per family:

- metrics/accuracy.rs   learning curve, category accuracy, temporal, calibration
- metrics/approvals.rs  false-approve/deny, counterfactuals, incidents, time-to-correct
- metrics/perf.rs       tier / latency / cache / distribution / impact / evolution

In-module `super::` rewritten to absolute `crate::brain::` so handlers survive
the move; compute_* + summary structs (TierStats/LatencySummary/CacheSummary/
compute_counterfactuals) re-exported so brain_screen + review keep resolving
`brain::metrics::*`. Full suite + clippy -D warnings green.
